Refactor prefix
