+++ src/commands/guild/prefix.ts
@@ -1,17 +1,15 @@
 import { Message } from "discord.js";
 import Keyv from "keyv";
-import { userPermissions } from "../../discordUserPermissions";
+import { userPermissions } from "../../models/discordUserPermissions";
 import { DATABASE } from "../../../config.json";
 import logger from "../../logging";
 
-const getGuildPrefix = async (message: Message, guildPrefixes: Keyv): Promise<string> => {
-  const guildId: string = message.guild.id;
+const getGuildPrefix = async (guildId: string, guildPrefixes: Keyv): Promise<string> => {
   const prefix: string = await guildPrefixes.get(guildId);
   return `Prefix is \`${prefix ?? "!"}\``;
 };
 
-const setGuildPrefix = async (message: Message, guildPrefixes: Keyv, newPrefix: string): Promise<string> => {
-  const guildId: string = message.guild.id;
+const setGuildPrefix = async (guildId: string, guildPrefixes: Keyv, newPrefix: string): Promise<string> => {
   await guildPrefixes.set(guildId, newPrefix);
   return `You changed the prefix to \`${newPrefix}\``;
 };
@@ -19,17 +17,17 @@ const setGuildPrefix = async (message: Message, guildPrefixes: Keyv, newPrefix: 
 const dispatchPrefixFun = async (message: Message, keyvGuildConfig: Keyv, args: string[]): Promise<string> => {
   let reply: string;
 
-  if (!args.length) {
-    reply = await getGuildPrefix(message, keyvGuildConfig);
+  if (!args.length || args.length === 0) {
+    reply = await getGuildPrefix(message.guild.id, keyvGuildConfig);
   } else if (args.length > 1) {
     reply = "You provided too many arguments. The prefix should be exactly one with a length of one. E.g.: `$`";
     throw new Error(reply);
-  } else if (args.length === 1) {
+  } else {
     if (args[0].length > 1) {
       reply = "The provided prefix is too long. Make sure it has a length of 1";
       throw new Error(reply);
     } else {
-      reply = await setGuildPrefix(message, keyvGuildConfig, args[0]);
+      reply = await setGuildPrefix(message.guild.id, keyvGuildConfig, args[0]);
     }
   }
   return reply;
@@ -45,7 +43,7 @@ module.exports = {
     const keyvGuildConfig: Keyv = new Keyv(DATABASE.CONNECTION_STRING, {
       namespace: "guildConfig",
     });
-    const reply = await dispatchPrefixFun(message, keyvGuildConfig, args).catch(e => {
+    const reply: string = await dispatchPrefixFun(message, keyvGuildConfig, args).catch(e => {
       logger.error(e);
       return e.message;
     });
