+++ src/models/discordUserPermissions.ts
